requirements.txt
djangorestframework
psycopg2
PyJWT
requests
drf-yasg
mozilla_django_oidc
markdown
djangorestframework-simplejwt
